# Optimize `String.repeat()` (3.x)