This backports the [optimization done in `master`](https://github.com/godotengine/godot/pull/64489).

**Testing project:** [test_string_repeat_3.x.zip](https://github.com/godotengine/godot/files/9440068/test_string_repeat_3.x.zip)
*Requires readding and exposing a `repeat_slow()` method with the old code to work.*

## Benchmark

**OS:** Fedora 36
**CPU:** Intel Core i7-6700K @ 4.4 GHz
**Compiler:** Clang 14.0.5

*Performed on an editor debug build, but a release build should also benefit significantly.*

Type | Before | After
-|-|-
len1 x4  |  5.967ms | 5.267ms
len1 x40 |  17.193ms | 5.445ms
len1 x400 |  126.042ms | 5.735ms
len1 x4000 | 1213.496ms | 7.698ms
len5 x4  |  10.426ms | 5.284ms
len5 x40 |  61.751ms | 5.6ms
len5 x400 | 569.243ms | 6.278ms
len45 x4 |  50.869ms | 5.38ms
len45 x40 | 462.394ms | 6.103ms
len250 x4 | 266.615ms | 5.668ms